voice_rec/requirements.txt
python-dotenv
uvicorn
fastapi
python-multipart
lameenc
dashscope